+++ src/main/kotlin/wishKnish/knishIO/client/mutation/MutationWithdrawBufferToken.kt
@@ -3,20 +3,15 @@
 package wishKnish.knishIO.client.mutation
 
 import wishKnish.knishIO.client.Molecule
-import wishKnish.knishIO.client.Wallet
 import wishKnish.knishIO.client.httpClient.HttpClient
 
 class MutationWithdrawBufferToken @JvmOverloads constructor(
   httpClient: HttpClient,
   override val molecule: Molecule? = null
 ) : MutationProposeMolecule(httpClient, molecule) {
-  @JvmOverloads
-  fun fillMolecule(
-    recipients: Map<String, Number>,
-    signingWallet: Wallet? = null
-  ) {
+  fun fillMolecule(recipients: Map<String, Number>) {
     molecule?.apply {
-      initWithdrawBuffer(recipients, signingWallet)
+      initWithdrawBuffer(recipients)
       sign()
       check(sourceWallet)
     }